+++ crates/masm-analysis/src/unconstrained_advice/mod.rs
@@ -1,6 +1,7 @@
 //! Interprocedural analysis for unconstrained advice reaching U32 and non-zero sinks.
 
 mod address;
+mod call_transfer;
 mod domain;
 mod grouping;
 mod inter;
